refactor: Home画面の再リファクタリング #175
