+++ front/src/app/HomePresentation.test.tsx
@@ -1,4 +1,4 @@
-import { HomePresentation } from "@/app/presentation/Home/HomePresentation";
+import { HomePresentation } from "@/app/HomePresentation";
 import { LinkButtonProps } from "@/types";
 import { render, screen, waitFor } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
